Modularised code, added Enums
Made code more modular by creating new functions `board_move_counter` and `df_combiner`. Moreover, made an Enum for Hans Niemanns color. Renamed boardMaker.py to board_maker.py.
